feat(consume): refactor the consume pytest plugin.
