+++ src/ethereum_test_tools/rpc/__init__.py
@@ -2,6 +2,6 @@
 JSON-RPC methods and helper functions for EEST consume based hive simulators.
 """
 
-from .rpc import BlockNumberType, EngineRPC, EthRPC
+from .rpc import BlockNumberType, DebugRPC, EngineRPC, EthRPC
 
-__all__ = ["EthRPC", "BlockNumberType", "EngineRPC"]
+__all__ = ["EthRPC", "BlockNumberType", "EngineRPC", "DebugRPC"]
