+++ src/pytest_plugins/consume/ini_files/pytest-consume-all.ini
@@ -2,14 +2,12 @@
 console_output_style = count
 minversion = 7.0
 python_files = test_*
-testpaths = tests_consume/test_direct.py tests_consume/test_via_rlp.py tests_consume/test_via_engine_api.py
+testpaths = src/pytest_plugins/consume/direct/test_direct.py src/pytest_plugins/consume/rlp/test_rlp.py src/pytest_plugins/consume/engine/test_engine.py
 addopts = 
     -rxXs
     --tb short
-    -p pytest_plugins.consume.consume
-    -p pytest_plugins.consume.direct
-    -p pytest_plugins.consume.rlp
-    -p pytest_plugins.consume.engine
-    -p pytest_plugins.consume.simulator_common
+    -p pytest_plugins.consume.common
+    -p pytest_plugins.consume.simulators.engine.engine
+    -p pytest_plugins.consume.simulators.simulator_common
     -p pytest_plugins.pytest_hive.pytest_hive
     -p pytest_plugins.test_help.test_help
