+++ src/pytest_plugins/consume/ini_files/pytest-consume-direct.ini
@@ -2,10 +2,10 @@
 console_output_style = count
 minversion = 7.0
 python_files = test_direct.py
-testpaths = tests_consume/test_direct.py
+testpaths = src/pytest_plugins/consume/direct/test_direct.py
 addopts =
     -rxXs
     --tb short    
-    -p pytest_plugins.consume.consume
-    -p pytest_plugins.consume.direct
+    -p pytest_plugins.consume.common
+    -p pytest_plugins.consume.direct.conftest
     -p pytest_plugins.test_help.test_help
