+++ src/pytest_plugins/consume/ini_files/pytest-consume-engine.ini
@@ -1,13 +1,12 @@
 [pytest]
 console_output_style = count
 minversion = 7.0
-python_files = test_via_engine_api.py
-testpaths = tests_consume/test_via_engine_api.py
+python_files = test_via_engine.py
+testpaths = src/pytest_plugins/consume/hive_simulators/engine/test_via_engine.py
 addopts = 
     -rxXs
     --tb short    
-    -p pytest_plugins.consume.consume
-    -p pytest_plugins.consume.engine
-    -p pytest_plugins.consume.simulator_common
+    -p pytest_plugins.consume.common
+    -p pytest_plugins.consume.hive_simulators.engine.conftest
     -p pytest_plugins.pytest_hive.pytest_hive
     -p pytest_plugins.test_help.test_help
