+++ src/pytest_plugins/consume/ini_files/pytest-consume-rlp.ini
@@ -2,12 +2,11 @@
 console_output_style = count
 minversion = 7.0
 python_files = test_via_rlp.py
-testpaths = tests_consume/test_via_rlp.py
+testpaths = src/pytest_plugins/consume/hive_simulators/rlp/test_via_rlp.py
 addopts = 
     -rxXs
     --tb short    
-    -p pytest_plugins.consume.consume
-    -p pytest_plugins.consume.rlp
-    -p pytest_plugins.consume.simulator_common    
+    -p pytest_plugins.consume.common
+    -p pytest_plugins.consume.hive_simulators.rlp.conftest
     -p pytest_plugins.pytest_hive.pytest_hive
     -p pytest_plugins.test_help.test_help
